// src/types/legalitasLahan.ts
export interface LegalitasLahan {
    nomorSTDB: string;
    idDataKebun: string;
    nama: string;
    tempatTanggalLahir: string;
    nik: string;
    alamat: string;
    lokasiKebun: string;
    statusKepemilikanLahan: string;
    nomorSertifikat: string;
    luasArealKebun: number;
    luasKebun: string;
    jenisTanaman: string;
    produksiPerHaPertahun: string;
    asalBenih: string;
    polaTanam: string;
    jenisPupuk: string;
    mitraPengolahan: string;
    jenisTanah: string;
    tahunTanam: string;
    usahaLainDikebun: string;
    cidFotoLahan: string;
    longLatitude: string;
    waktuPengajuanSTDB: string;
    niKoperasi: string;
    namaKoperasi: string;
    statusKonfirmasiKoperasi: string;
    waktuStatusKonfirmasiKoperasi: string;
    alasanPenolakan: string;
    nipDinas: string;
    namaDinas: string;
    statusPenerbitanSTDB: string;
    statusPenerbitLegalitas: string;
    waktuPenerbitanSTDB: string;
    waktuPemetaanKebun: string;
    waktuUpdatePemetaanKebun: string;
    createdAt: string;
    statusPemetaan: string;
    namaKonfirmator: string;
    waktuKonfirmator: string;
    cidFotoKebun: string;
    alamatKebun: string;
    idPemetaanKebun: string;
    waktuPengajuan: string;
    statusKonfirmator: string;
    cidFileLegalitasLahan: string;
    statusVerifikator: string;
    nipVerifikator: string;
    waktuVerifikator: string;
    nipPenerbitLegalitas: string;
    cidFileLegalitasKebun: string;
    nikSurveryor: string;
    nikKonfirmator: string;
    pesanKonfirmator: string;
    statusKawasan: string;
    longitude: string;
    latitude: string;
    jumlahPohon: string;
    nikSurveyor: string;
    namaSurveyor: string;
    namaVerifikator: string;
    pesanVerifikator: string;
    WaktuPemetaanKebun: string;
    namaPenerbitLegalitas: string;
    waktuPenerbitLegalitas: string;
    updateWaktuPenerbitLegalitas: string;
    updateWaktuKonfirmator: string;
    waktuUpdateVerifikator: string;
}